% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_linters.R
\name{example_linters}
\alias{example_linters}
\alias{no_t_f_linter}
\alias{no_attach_linter}
\alias{prefer_arrow_assignment_linter}
\alias{no_one_length_linter}
\alias{no_sapply_linter}
\title{Example Linters}
\usage{
no_t_f_linter()

no_attach_linter()

prefer_arrow_assignment_linter()

no_one_length_linter()

no_sapply_linter()
}
\description{
A collection of example linters that demonstrate common patterns.
These can be used as-is or as templates for your own linters.
}
\section{Functions}{
\itemize{
\item \code{no_t_f_linter()}: Warns against using T/F instead of TRUE/FALSE

\item \code{no_attach_linter()}: Warns against using attach()

\item \code{prefer_arrow_assignment_linter()}: Suggests using <- instead of = for assignment

\item \code{no_one_length_linter()}: Warns against 1:length(x) pattern

\item \code{no_sapply_linter()}: Style guide: warns against sapply

}}
