% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_linter.R
\name{create_function_call_linter}
\alias{create_function_call_linter}
\title{Create a Function Call Linter}
\usage{
create_function_call_linter(
  function_names,
  message,
  linter_name,
  type = c("warning", "error", "style")
)
}
\arguments{
\item{function_names}{Character vector of function names to lint against.}

\item{message}{A character string with the lint message. Use {function} as a
placeholder for the actual function name found.}

\item{linter_name}{A character string naming the linter.}

\item{type}{The type of lint. Defaults to "warning".}
}
\value{
A linter function.
}
\description{
Specialized helper for creating linters that target specific function calls.
This is a common pattern in linting, so this function makes it even easier.
}
\examples{
\dontrun{
# Warn against using paste0 in favor of paste
no_paste0_linter <- create_function_call_linter(
  function_names = "paste0",
  message = "Use paste(..., sep = '') instead of {function}.",
  linter_name = "no_paste0_linter"
)
}

}
