% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_linters.R
\name{limit_line_length}
\alias{limit_line_length}
\title{Limit Line Length}
\usage{
limit_line_length(
  max_length = 80,
  message = NULL,
  type = c("style", "warning", "error")
)
}
\arguments{
\item{max_length}{Maximum allowed line length. Default is 80.}

\item{message}{The lint message.}

\item{type}{The lint type. Defaults to "style".}
}
\value{
A linter function.
}
\description{
Flag lines that exceed a certain character length.
}
\examples{
\dontrun{
# Standard 80 characters
line_length_80 <- limit_line_length(80)

# More strict
line_length_72 <- limit_line_length(72)
}

}
